+++ Editor/Launchers/GVim.cs
@@ -7,12 +7,19 @@ namespace EasyEditor.Launchers
     using UnityEditor;
     using UnityEngine;
 
-    internal class GVim : ILauncher
+    internal class GVim : Launcher
     {
-        private static readonly string PrefPrefix = $"{typeof(GVim).FullName}";
-        private static readonly string PrefRestartOmniSharp = $"{PrefPrefix}.RestartOmniSharp";
+        [InitializeOnLoad]
+        private static class Settings
+        {
+            public static readonly Setting restartOmniSharp = new Setting(
+                "RestartOmniSharp",
+                "Restart OmniSharp server on sync",
+                "Upon synchronization, send a command (--remote-send) to the gVim server to trigger a OmniSharp server reload. omnisharp-vim required.",
+                "GVim");
+        }
 
-        public CodeEditor.Installation[] Installations => new CodeEditor.Installation[]
+        public override CodeEditor.Installation[] Installations => new CodeEditor.Installation[]
         {
 #if UNITY_EDITOR_WIN
             new CodeEditor.Installation()
@@ -28,7 +35,7 @@ namespace EasyEditor.Launchers
 #endif
         };
 
-        public bool Launch(string editorPath, LaunchDescriptor descriptor)
+        public override bool Launch(string editorPath, LaunchDescriptor descriptor)
         {
             string args = string.Format("--servername \"{0}\" --remote-silent +$(Line) $(File)", descriptor.ProjectName);
 
@@ -51,7 +58,7 @@ namespace EasyEditor.Launchers
 
         private void ReloadOmniSharpServer()
         {
-            string args = string.Format("--servername \"{0}\" --remote-send '<C-\\><C-N>:silent OmniSharpRestartServer<CR>'", Preferences.ProjectName);
+            string args = string.Format("--servername \"{0}\" --remote-send '<C-\\><C-N>:silent OmniSharpRestartServer<CR>'", Preferences.projectName);
 
             Process process = new Process
             {
@@ -68,23 +75,23 @@ namespace EasyEditor.Launchers
             _ = process.Start();
         }
 
-        public void SyncAll()
+        public override void SyncAll()
         {
-            if (Preferences.IsActive && EditorPrefs.GetBool(PrefRestartOmniSharp))
+            if (Preferences.IsActive && Settings.restartOmniSharp.GetBool())
             {
                 ReloadOmniSharpServer();
             }
         }
 
-        public void SyncIfNeeded(string[] addedFiles, string[] deletedFiles, string[] movedFiles, string[] movedFromFiles, string[] importedFiles)
+        public override void SyncIfNeeded(string[] addedFiles, string[] deletedFiles, string[] movedFiles, string[] movedFromFiles, string[] importedFiles)
         {
-            if (Preferences.IsActive && EditorPrefs.GetBool(PrefRestartOmniSharp))
+            if (Preferences.IsActive && Settings.restartOmniSharp.GetBool())
             {
                 ReloadOmniSharpServer();
             }
         }
 
-        public bool MatchesExecutable(string editorPath)
+        public override bool MatchesExecutable(string editorPath)
         {
             string vimPath = Path.Combine(Path.GetDirectoryName(editorPath), "vim.exe");
             Process process = new Process
@@ -109,16 +116,15 @@ namespace EasyEditor.Launchers
             return output.StartsWith("VIM - Vi IMproved");
         }
 
-        public void OnGUI()
+        public override void OnGUI()
         {
             EditorGUI.BeginChangeCheck();
-            GUIContent restartOmniSharpContent = new GUIContent(
-                "Restart OmniSharp server on sync",
-                "Upon synchronization, send a command (--remote-send) to the gVim server to trigger a OmniSharp server reload. omnisharp-vim required.");
-            bool b = EditorGUILayout.Toggle(restartOmniSharpContent, EditorPrefs.GetBool(PrefRestartOmniSharp));
+            Setting s = Settings.restartOmniSharp;
+            GUIContent c = new GUIContent(s.description, s.tooltip);
+            bool b = EditorGUILayout.Toggle(c, s.GetBool());
             if (EditorGUI.EndChangeCheck())
             {
-                EditorPrefs.SetBool(PrefRestartOmniSharp, b);
+                s.SetBool(b);
                 Event.current.Use();
             }
         }